pyproject.toml
[project]
name = "dls-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.12",
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "google-api-core>=2.25.0",
    "google-genai>=1.18.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.5",
    "langchain-huggingface>=0.2.0",
    "langchain-weaviate>=0.0.5",
    "onnx>=1.18.0",
    "onnxruntime>=1.22.0",
    "optimum[exporters]>=1.25.3",
    "pdf2img>=0.1.2",
    "pdfplumber>=0.11.6",
    "pillow>=11.2.1",
    "pymupdf>=1.26.0",
    "pypdf>=5.6.0",
    "pytesseract>=0.3.13",
    "requests>=2.32.3",
    "sentence-transformers>=4.1.0",
    "tenacity>=9.1.2",
    "weaviate-client>=4.15.0",
]
